rpi-update: Allow some explicit non-custom kernels (#20)
rpi-update tries to avoid rendering a Pi non-bootable by detecting
when a custom kernel is being used. Since dtb filenames are not
qualified with the kernel name, overwriting the dtb files for a
custom kernel but leaving the custom kernel unchanged may stop the
Pi booting.

However, adding kernel=kernel8.img to config.txt is a way to select a
kernel with 4kB pages on a Pi 5. Treat kernel8.img and kernel_2712.img
as non-custom kernels, and also allow SKIP_KERNEL=0 to be used to
force the kernel to be updated.